requirements.txt
fastapi>=0.110
python-dotenv>=1.0
uvicorn>=0.27
pydantic>=2.5
jinja2>=3.1
google-generativeai>=0.8.6